changed logic for finding previous event; using similar system to oystedal
we look for the very previous event that has the same thread ID as the new event, except for cases where the event type is 2. In those cases, we look for the previous trace ID with type 1 that comes before the current trace ID. We don't delete previous records of type 1 because they might be needed by several cores, and we don't want to keep in the config file the number of cores used.