moved rpi traces to a subfolder to improve sorting in analysis GUI
the problem is that traces created on the laptop appear much more recent than the rpi traces, most likely because the RPI clock is out of sync. This means that we have trouble finding the most recent rpi trace. We fixed this in the analysis tool by using kivy's FileChooser data structure instead of an ad-hoc file chooser solution